Pre-empting std replacements on bitcoin-core
There are overrides for `serialise.h` in bitcoin-core, to replace the
use of exception with chromium's CHECK calls. The preprocessor
instructions are very tricky, and sketchy, and due to a recent libcxx
update, build failures started occurring when such replacements started
affecting standard headers.

This change preempts the inclusion of such headers on the overriding
file itself, in order to avoid substitution in standard library files.