[ios] Removing unused `chromium_find_in_page` provider
This provider seems to be broken in upstream at the moment, as
`chromium_find_in_page` has as a dependency `find_in_page:features`,
which seems to be non-existent.

The provider is being removed regardless, as it is not in use, and
therfore it is merely adding to the build/linking time.