Deactivate dangling pointer checks
Upstream has enabled dangling pointer checks as default for debug linux
builds. Brave's codebase has not been checked for these fixes yet. This
change is patch to make sure that for now these checks are not enabled
for debug mode in linux, as they won't work in the meantime.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/4a76393f4f5e6eae945409785e9bb03080241d24

commit 4a76393f4f5e6eae945409785e9bb03080241d24
Author: Paul Semel <paulsemel@chromium.org>
Date:   Fri Jul 28 13:36:39 2023 +0000

    dpd: enable dpd on non official builds + dcheck_always_on

    Similarly to enabling dpd on debug builds, we want to broaden the usage
    of DPD by developers. With this CL, we enable DPD on non official builds
    that have dcheck_always_on activated, which is a common configuration
    for chromium developers. This should also ensure we are not affecting
    official builds.